refactor: アカウント設定とプロフィール関連のファイルをfeaturesディレクトリに集約
